pico-sdk: Unify SRCREV for the Pico SDK
Both the examples and the native tools are built using the Pico SDK,
hence we should be using the same commit ID to build both.

Use the PICO_SDK_SRCREV variable to unify the commit used for the Pico
SDK across the layer (whether the layer.conf is the proper place for
this variable is a separate matter but it should do for now).

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro@enedino.org>