pico-sdk: Switch SDK approach
At first, we attempted to pull in the pico-sdk by following the approach used by
meta-freertos, the rpi-pico-sdk class added an extra SRC_URI along with the
required wiring for it, this works fine for meta-freertos because it pulls in
the kernel sources that way (used to build).
However, in the SDK case, its sources arent being built but are instead used by
the buildsystem to perform the build, hence for this scenario it makes more
sense to use the approach used by meta-zephyr instead, by creating a pico-sdk
package which we can DEPEND on to be able to use it for our applications.

This also simplifies SRC_URI and SRCREV usage, also, the rpi-pico-sdk class is
still used to automatically add dependencies and wire the SDK properly.

Tested on both freertos-demo-pico and pico-examples.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandro@enedino.org>